+++ apps/atr-one/app/frontend/styles/tailwind-directives.css
@@ -1,0 +1,3 @@
+@tailwind base;
+@tailwind components;
+@tailwind utilities;
