+++ apps/atr-one/app/views/layouts/visualization.html.slim
@@ -5,7 +5,7 @@ html
     meta name="viewport" content="width=device-width,initial-scale=1"
 
     = vite_client_tag
-    = vite_javascript_tag 'application'
+    = vite_javascript_tag 'visualization'
 
     = csrf_meta_tags
